Extracted base Azure Storage logic into separate
