+++ src/Infrastructure/VolleyM.Infrastructure.AzureStorage/AzureTableStorageOptions.cs
@@ -1,0 +1,7 @@
+﻿namespace VolleyM.Infrastructure.AzureStorage
+{
+    public class AzureTableStorageOptions
+    {
+        public string ConnectionString { get; set; }
+    }
+}
