+++ src/Infrastructure/VolleyM.Infrastructure.IdentityAndAccess.AzureStorage/TableConfiguration/IdentityContextTableStorageOptions.cs
@@ -1,9 +1,9 @@
-﻿namespace VolleyM.Infrastructure.IdentityAndAccess.AzureStorage.TableConfiguration
-{
-    public class IdentityContextTableStorageOptions
-    {
-        public string ConnectionString { get; set; }
+﻿using VolleyM.Infrastructure.AzureStorage;
 
+namespace VolleyM.Infrastructure.IdentityAndAccess.AzureStorage.TableConfiguration
+{
+    public class IdentityContextTableStorageOptions : AzureTableStorageOptions
+    {
         public string UsersTable { get; set; } = "users";
     }
 }
